rename MDT
