+++ src/components/pages/GenericMdt/__tests__/GenericMdtCreateForm.test.js
@@ -3,7 +3,7 @@ import Enzyme, { shallow } from 'enzyme';
 import Adapter from 'enzyme-adapter-react-15';
 import configureStore from 'redux-mock-store';
 
-import MDTsCreateForm from '../MDTsCreate/MDTsCreateForm';
+import MDTsCreateForm from '../GenericMdtCreate/GenericMdtCreateForm';
 import { valuesNames, valuesLabels } from '../forms.config';
 
 Enzyme.configure({ adapter: new Adapter() });
