+++ src/components/pages/GenericMdt/ducks/fetch-patient-generic-mdt-create.duck.js
@@ -1,8 +1,7 @@
 import { ajax } from 'rxjs/observable/dom/ajax';
 import { createAction } from 'redux-actions';
-
-import { usersUrls } from '../../../../config/server-urls.constants'
-import { fetchPatientMDTsRequest } from './fetch-patient-mdts.duck'
+import { usersUrls } from '../../../../config/server-urls.constants';
+import { fetchPatientMDTsRequest } from './fetch-patient-generic-mdt.duck';
 
 export const FETCH_PATIENT_MDTS_CREATE_REQUEST = 'FETCH_PATIENT_MDTS_CREATE_REQUEST';
 export const FETCH_PATIENT_MDTS_CREATE_SUCCESS = 'FETCH_PATIENT_MDTS_CREATE_SUCCESS';
