+++ src/components/pages/GenericMdt/ducks/fetch-patient-generic-mdt-detail-edit.duck.js
@@ -2,7 +2,7 @@ import { ajax } from 'rxjs/observable/dom/ajax';
 import { createAction } from 'redux-actions';
 
 import { usersUrls } from '../../../../config/server-urls.constants'
-import { fetchPatientMDTsUpdateRequest } from './fetch-patient-mdts.duck'
+import { fetchPatientMDTsUpdateRequest } from './fetch-patient-generic-mdt.duck'
 
 export const FETCH_PATIENT_MDTS_DETAIL_EDIT_REQUEST = 'FETCH_PATIENT_MDTS_DETAIL_EDIT_REQUEST';
 export const FETCH_PATIENT_MDTS_DETAIL_EDIT_SUCCESS = 'FETCH_PATIENT_MDTS_DETAIL_EDIT_SUCCESS';
