+++ src/components/pages/TransfersOfCare/__tests__/TransfersOfCare.test.js
@@ -44,7 +44,7 @@ const storeWithFormsError = mockStore(Object.assign({
           [valuesNames.RECORDS_SOURCE]: 'ethercis',
           [valuesNames.SOURCE_ID]: '0318e94e-1803-46fe-a92a-647eca029323',
           [valuesNames.TYPE]: 'diagnosis',
-          [valuesNames.RECORDS_TYPE]: 'Problems / Diagnosis',
+          [valuesNames.RECORDS_TYPE]: 'Diagnosis',
         },
         {
           [valuesNames.RECORDS_DATE]: '11-Feb-2015',
@@ -93,7 +93,7 @@ const storeWithDetail = mockStore(Object.assign({
           [valuesNames.RECORDS_SOURCE]: 'ethercis',
           [valuesNames.SOURCE_ID]: '0318e94e-1803-46fe-a92a-647eca029323',
           [valuesNames.TYPE]: 'diagnosis',
-          [valuesNames.RECORDS_TYPE]: 'Problems / Diagnosis',
+          [valuesNames.RECORDS_TYPE]: 'Diagnosis',
         },
         {
           [valuesNames.RECORDS_DATE]: '11-Feb-2015',
