+++ src/config/for-test.constants.js
@@ -1,6 +1,6 @@
 export const testConstants = {
   hostName: 'http://dev.ripple.foundation:8000',
-  token: 'eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJleHAiOjE1MjkwNDk1MjMsImlhdCI6MTUyOTA0ODkyMywiaXNzIjoicWV3ZC5qd3QiLCJhcHBsaWNhdGlvbiI6InJpcHBsZS1jZHItb3BlbmVociIsInRpbWVvdXQiOjYwMCwicWV3ZCI6ImQzOTAxY2FjNzczMmJlZjQ3NjY5MmY3NDFlNzNiNjYxYzI3NDFmNDMzZGQ4YmVmMTA3NDZjZWUxMzExMjQzYTQyNjE0YjE5ZDdhYzU1NjQ4MmNmZjFkNDEyNzU0YmRlYWNhNWNmYWUzZDFiNTc5MWExOGNhMzVhNTYzN2MwOTcyNDc2M2YzZDQ2NTlkNGUwZGViZDE5MThmNGEzYzFhNzViNmMxZGEzOTIwYmRjYzY3M2NjNWNiYzgyN2QxNGRiZiIsIm5oc051bWJlciI6Ijk5OTk5OTkwMDAiLCJlbWFpbCI6Iml2b3IuY294QHJpcHBsZS5mb3VuZGF0aW9uIiwicm9sZSI6InBoclVzZXIiLCJ1aWQiOiJlYTY3NTU2Yy1mZTU3LTRjZjAtYmFhMy1mNjRlYzNjZjRiYzAiLCJvcGVuaWQiOnsianRpIjoiOWZlNjhiMDEtNWI2ZC00ZDNkLWEyOGYtMmI2OGZjZThmOGYwIiwiZXhwIjoxNTI5MDQ5MjE2LCJuYmYiOjAsImlhdCI6MTUyOTA0ODkxNiwiaXNzIjoiaHR0cHM6Ly9rZXljbG9hay5kZXYxLnNpZ25pbi5uaHMudWsvY2ljYXV0aC9yZWFsbXMvTkhTIiwiYXVkIjoibGVlZHMtaGVsbSIsInN1YiI6IjcxYmE1MGYyLWI0M2ItNGVhNC1iZTBmLWY1YjdhZWJhMjlhNSIsInR5cCI6IklEIiwiYXpwIjoibGVlZHMtaGVsbSIsImF1dGhfdGltZSI6MTUyOTA0ODkxNiwic2Vzc2lvbl9zdGF0ZSI6ImVhNjc1NTZjLWZlNTctNGNmMC1iYWEzLWY2NGVjM2NmNGJjMCIsImFjciI6IjEiLCJwcmVmZXJyZWRfdXNlcm5hbWUiOiJpdm9yLmNveEByaXBwbGUuZm91bmRhdGlvbiIsImVtYWlsIjoiaXZvci5jb3hAcmlwcGxlLmZvdW5kYXRpb24iLCJuaHNOdW1iZXIiOiI5OTk5OTk5MDAwIiwiaWRfdG9rZW4iOiJleUpoYkdjaU9pSlNVekkxTmlJc0luUjVjQ0lnT2lBaVNsZFVJaXdpYTJsa0lpQTZJQ0kzV0dsTmRIVkdTSE4wTVZkdFlVcHBTWGRKWVZaSlpsSk9helE1U3psVE5YbFdhMHRoTWtaWlpIQTBJbjAuZXlKcWRHa2lPaUk1Wm1VMk9HSXdNUzAxWWpaa0xUUmtNMlF0WVRJNFppMHlZalk0Wm1ObE9HWTRaakFpTENKbGVIQWlPakUxTWprd05Ea3lNVFlzSW01aVppSTZNQ3dpYVdGMElqb3hOVEk1TURRNE9URTJMQ0pwYzNNaU9pSm9kSFJ3Y3pvdkwydGxlV05zYjJGckxtUmxkakV1YzJsbmJtbHVMbTVvY3k1MWF5OWphV05oZFhSb0wzSmxZV3h0Y3k5T1NGTWlMQ0poZFdRaU9pSnNaV1ZrY3kxb1pXeHRJaXdpYzNWaUlqb2lOekZpWVRVd1pqSXRZalF6WWkwMFpXRTBMV0psTUdZdFpqVmlOMkZsWW1FeU9XRTFJaXdpZEhsd0lqb2lTVVFpTENKaGVuQWlPaUpzWldWa2N5MW9aV3h0SWl3aVlYVjBhRjkwYVcxbElqb3hOVEk1TURRNE9URTJMQ0p6WlhOemFXOXVYM04wWVhSbElqb2laV0UyTnpVMU5tTXRabVUxTnkwMFkyWXdMV0poWVRNdFpqWTBaV016WTJZMFltTXdJaXdpWVdOeUlqb2lNU0lzSW5CeVpXWmxjbkpsWkY5MWMyVnlibUZ0WlNJNkltbDJiM0l1WTI5NFFISnBjSEJzWlM1bWIzVnVaR0YwYVc5dUlpd2laVzFoYVd3aU9pSnBkbTl5TG1OdmVFQnlhWEJ3YkdVdVptOTFibVJoZEdsdmJpSXNJbTVvYzA1MWJXSmxjaUk2SWprNU9UazVPVGt3TURBaWZRLm1wSHdvM2pwU3ZTQW1aR085bjYzbDl5UV9oNFdsZ3FMYlRVRFVlNGtSVk93M0JfalgxRklEVTd2Qi1HQ2VFRDBybU9sUDZza1hhQ3UyNldXQ2lpVTNNWVY4Ml9pc1F3YVJkdmZVSS1tb0R5VFJIeXpjNm80WEN0SUZUWEhZWWtnT2Z4T0FhOFUtZVJWYkYtNVB3czZ1NXlUWEwxbWp4U2VJSEk4NVJNaGhyS2N1a3JhTXFBcy15aVhWNGM1THVBOEM4S0U2QVNid1NvNGNfTm9Gd0VuV3JsaUxMQVBTTGx1ZHhad0NCUVdHQTRHYXc0RFh0RHVOdktnQnJrX1hpLXgwQWtOUzMyV0pIVzN5dXNBV29MRHZ5a3ZtaXFZVi1xRWNuOFhjVFBhaHk1NUpHZkJWN21BY2xNMjZUb3doRTRMSmlvVVF3Z2Y1NEdOMUM3dVpyTm1CZyJ9fQ.CV8BVqMHd8m3GYzcFL-Ss2nr-jbeKQkfOugK3_8fg-g'
+  token: 'eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJleHAiOjE1MjkwNTUxODgsImlhdCI6MTUyOTA1NDU4OCwiaXNzIjoicWV3ZC5qd3QiLCJhcHBsaWNhdGlvbiI6InJpcHBsZS1jZHItb3BlbmVociIsInRpbWVvdXQiOjYwMCwicWV3ZCI6ImQzOTAxY2FjNzczMmJlZjQ3NjY5MmY3NDFlNzNiNjYxYzI3NDFmNDMzZGQ4YmVmMTA3NDZjZWUxMzExMjQzYTQyNjE0YjE5ZDdhYzU1NjQ4MmNmZjFkNDEyNzU0YmRlYWNhNWNmYWUzZDFiNTc5MWExOGNhMzVhNTYzN2MwOTcyNDc2M2YzZDQ2NTlkNGUwZGViZDE5MThmNGEzYzFhNzViNmMxZGEzOTIwYmRjYzY3M2NjNWNiYzgyN2QxNGRiZiIsIm5oc051bWJlciI6Ijk5OTk5OTkwMDAiLCJlbWFpbCI6Iml2b3IuY294QHJpcHBsZS5mb3VuZGF0aW9uIiwicm9sZSI6InBoclVzZXIiLCJ1aWQiOiJkMmIzNjUzNy1iOWVjLTRmMTgtYWY0Yy0zYTAzNGU2OWIzNDMiLCJvcGVuaWQiOnsianRpIjoiMDlhZDUxYmYtOTM4My00MzgzLThlZGEtZTNhYzQxZjU0ZmY5IiwiZXhwIjoxNTI5MDU0ODcwLCJuYmYiOjAsImlhdCI6MTUyOTA1NDU3MCwiaXNzIjoiaHR0cHM6Ly9rZXljbG9hay5kZXYxLnNpZ25pbi5uaHMudWsvY2ljYXV0aC9yZWFsbXMvTkhTIiwiYXVkIjoibGVlZHMtaGVsbSIsInN1YiI6IjcxYmE1MGYyLWI0M2ItNGVhNC1iZTBmLWY1YjdhZWJhMjlhNSIsInR5cCI6IklEIiwiYXpwIjoibGVlZHMtaGVsbSIsImF1dGhfdGltZSI6MTUyOTA1NDU3MCwic2Vzc2lvbl9zdGF0ZSI6ImQyYjM2NTM3LWI5ZWMtNGYxOC1hZjRjLTNhMDM0ZTY5YjM0MyIsImFjciI6IjEiLCJwcmVmZXJyZWRfdXNlcm5hbWUiOiJpdm9yLmNveEByaXBwbGUuZm91bmRhdGlvbiIsImVtYWlsIjoiaXZvci5jb3hAcmlwcGxlLmZvdW5kYXRpb24iLCJuaHNOdW1iZXIiOiI5OTk5OTk5MDAwIiwiaWRfdG9rZW4iOiJleUpoYkdjaU9pSlNVekkxTmlJc0luUjVjQ0lnT2lBaVNsZFVJaXdpYTJsa0lpQTZJQ0kzV0dsTmRIVkdTSE4wTVZkdFlVcHBTWGRKWVZaSlpsSk9helE1U3psVE5YbFdhMHRoTWtaWlpIQTBJbjAuZXlKcWRHa2lPaUl3T1dGa05URmlaaTA1TXpnekxUUXpPRE10T0dWa1lTMWxNMkZqTkRGbU5UUm1aamtpTENKbGVIQWlPakUxTWprd05UUTROekFzSW01aVppSTZNQ3dpYVdGMElqb3hOVEk1TURVME5UY3dMQ0pwYzNNaU9pSm9kSFJ3Y3pvdkwydGxlV05zYjJGckxtUmxkakV1YzJsbmJtbHVMbTVvY3k1MWF5OWphV05oZFhSb0wzSmxZV3h0Y3k5T1NGTWlMQ0poZFdRaU9pSnNaV1ZrY3kxb1pXeHRJaXdpYzNWaUlqb2lOekZpWVRVd1pqSXRZalF6WWkwMFpXRTBMV0psTUdZdFpqVmlOMkZsWW1FeU9XRTFJaXdpZEhsd0lqb2lTVVFpTENKaGVuQWlPaUpzWldWa2N5MW9aV3h0SWl3aVlYVjBhRjkwYVcxbElqb3hOVEk1TURVME5UY3dMQ0p6WlhOemFXOXVYM04wWVhSbElqb2laREppTXpZMU16Y3RZamxsWXkwMFpqRTRMV0ZtTkdNdE0yRXdNelJsTmpsaU16UXpJaXdpWVdOeUlqb2lNU0lzSW5CeVpXWmxjbkpsWkY5MWMyVnlibUZ0WlNJNkltbDJiM0l1WTI5NFFISnBjSEJzWlM1bWIzVnVaR0YwYVc5dUlpd2laVzFoYVd3aU9pSnBkbTl5TG1OdmVFQnlhWEJ3YkdVdVptOTFibVJoZEdsdmJpSXNJbTVvYzA1MWJXSmxjaUk2SWprNU9UazVPVGt3TURBaWZRLkZJUnVQLVpyN01meUZEc2RhbFpLUlBhMEs3V0FvUVBxTU90c3FSdHNnV3ZURXZ5OVFYNF9fajBwanpMMWtjejVIcVBMN0Y4X0NuNjNFWjBVSkJnRDFyT0Z3WmpfNDc2Q3hpS0RlWUhycFpOR0ZPc2s5ZF93anN6Qi1MU2p2bUs3VE5OOTJzcFNmNmVOTlFUS0ZsVll3cUF2a1ZOQi1hSTVLTjJzNzY0U19Id01xNW9kM1FZcFpxVWF6NFVjZVF0ZWJkdi1lcU1Zb0xWOFowdGFzRjVFVGU4Y3dTbWZIT2NJR05OZEN0d2lqUlJabG92TjluUG5XOUQxbWJSRnVrbUlTZFhnR3UxSU5uUDVkZENCS1VqalpHMHl4RWx3RTZJYnBsQ0wtRVp5S3M1Ym5odVlmWEtkNm5rc2JXcVRNaG5zQlQxQjhnd2xVVkFjcWJ0aGVPTjRfZyJ9fQ.qMPo4yQ6BhX97K3DS0tdfQYXALdHdJGho5WKeLkIqrI'
 };
 
 export const isDevMode = false;
